Update stride to 4.0.0.1-beta02-0926 & rename csproj
