+++ StrideLiveEditor/ComponentPropertyItem.cs
