+++ StrideLiveEditor/DataTypeEditors/BaseEditor.cs
