+++ StrideLiveEditor/DataTypeEditors/BooleanEditor.xaml
