+++ StrideLiveEditor/DataTypeEditors/BooleanEditor.xaml.cs
