+++ StrideLiveEditor/DataTypeEditors/EnumEditor.xaml
