+++ StrideLiveEditor/DataTypeEditors/EnumEditor.xaml.cs
