+++ StrideLiveEditor/DataTypeEditors/Int32Editor.xaml
