+++ StrideLiveEditor/DataTypeEditors/QuaternionEditor.xaml
