+++ StrideLiveEditor/DataTypeEditors/QuaternionEditor.xaml.cs
