+++ StrideLiveEditor/DataTypeEditors/RotationEditor.xaml
