+++ StrideLiveEditor/DataTypeEditors/RotationEditor.xaml.cs
