+++ StrideLiveEditor/DataTypeEditors/SingleEditor.xaml
