+++ StrideLiveEditor/DataTypeEditors/SingleEditor.xaml.cs
