+++ StrideLiveEditor/DataTypeEditors/UnsupportedEditor.xaml.cs
