+++ StrideLiveEditor/DataTypeEditors/Vector2Editor.xaml
