+++ StrideLiveEditor/DataTypeEditors/Vector3Editor.xaml
