+++ StrideLiveEditor/EntityComponentExpander.xaml
