+++ StrideLiveEditor/EntityComponentExpander.xaml.cs
