+++ StrideLiveEditor/EntityComponentInfo.cs
