+++ StrideLiveEditor/LiveEditorMainWindow.xaml
