+++ StrideLiveEditor/LogLevel.cs
