--- XenkoLiveEditor/StrideLiveEditor.csproj
@@ -1,33 +1,0 @@
-﻿<Project Sdk="MSBuild.Sdk.Extras/2.0.24">
-  <PropertyGroup>
-    <TargetFramework>net472</TargetFramework>
-    <PackageId>$(AssemblyName)</PackageId>
-    <Description>$(AssemblyName)</Description>
-    <StrideVersion Condition="'$(StrideVersion)' == ''">4.0.0.1-vvvv-0965</StrideVersion>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="PresentationCore" />
-    <Reference Include="PresentationFramework" />
-    <Reference Include="WindowsBase" />
-    <Reference Include="System.Xaml" />
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Update="Properties\Resources.resx" Generator="ResXFileCodeGenerator" LastGenOutput="Resources.Designer.cs" />
-    <Compile Update="Properties\Resources.Designer.cs" DesignTime="True" AutoGen="True" DependentUpon="Resources.resx" />
-    <Page Include="**\*.xaml" SubType="Designer" Generator="MSBuild:Compile" />
-    <Compile Update="**\*.xaml.cs" SubType="Designer" DependentUpon="%(Filename)" />
-    <Resource Include="Fonts\*.otf" />
-    <Resource Include="Images\*.png" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="MahApps.Metro" Version="1.6.5" />
-    <PackageReference Include="Nerdbank.GitVersioning" Version="2.3.186">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Stride.Engine" Version="$(StrideVersion)" />
-  </ItemGroup>
-</Project>
