+++ XenkoLiveEditorContainerProject.sln
@@ -1,7 +1,6 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.645
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30011.22
 MinimumVisualStudioVersion = 14.0.23107.0
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "XenkoLiveEditorContainerProject", "XenkoLiveEditorContainerProject", "{5523F72A-4E82-42D8-AA47-602BFFCCD8DC}"
 EndProject
@@ -9,7 +8,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "XenkoLiveEditorContainerPro
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "XenkoLiveEditorContainerProject.Game", "XenkoLiveEditorContainerProject\XenkoLiveEditorContainerProject.Game\XenkoLiveEditorContainerProject.Game.csproj", "{B4C976B6-4FF9-435A-BFAD-77D431EBE63F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XenkoLiveEditor", "XenkoLiveEditor\XenkoLiveEditor.csproj", "{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "StrideLiveEditor", "StrideLiveEditor\StrideLiveEditor.csproj", "{948AD18F-36C4-479F-B96E-0DD192D853C6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -81,30 +80,30 @@ Global
 		{B4C976B6-4FF9-435A-BFAD-77D431EBE63F}.Testing|Mixed Platforms.Build.0 = Release|Any CPU
 		{B4C976B6-4FF9-435A-BFAD-77D431EBE63F}.Testing|Windows.ActiveCfg = Release|Any CPU
 		{B4C976B6-4FF9-435A-BFAD-77D431EBE63F}.Testing|Windows.Build.0 = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.AppStore|Any CPU.ActiveCfg = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.AppStore|Any CPU.Build.0 = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.AppStore|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.AppStore|Mixed Platforms.Build.0 = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.AppStore|Windows.ActiveCfg = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.AppStore|Windows.Build.0 = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Debug|Windows.ActiveCfg = Debug|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Debug|Windows.Build.0 = Debug|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Release|Windows.ActiveCfg = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Release|Windows.Build.0 = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Testing|Any CPU.ActiveCfg = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Testing|Any CPU.Build.0 = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Testing|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Testing|Mixed Platforms.Build.0 = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Testing|Windows.ActiveCfg = Release|Any CPU
-		{E8B8FC87-95CB-41CC-91E1-6EBF964A0B7E}.Testing|Windows.Build.0 = Release|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.AppStore|Any CPU.ActiveCfg = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.AppStore|Any CPU.Build.0 = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.AppStore|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.AppStore|Mixed Platforms.Build.0 = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.AppStore|Windows.ActiveCfg = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.AppStore|Windows.Build.0 = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Debug|Windows.ActiveCfg = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Debug|Windows.Build.0 = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Release|Windows.ActiveCfg = Release|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Release|Windows.Build.0 = Release|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Testing|Any CPU.ActiveCfg = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Testing|Any CPU.Build.0 = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Testing|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Testing|Mixed Platforms.Build.0 = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Testing|Windows.ActiveCfg = Debug|Any CPU
+		{948AD18F-36C4-479F-B96E-0DD192D853C6}.Testing|Windows.Build.0 = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
