+++ XenkoLiveEditorContainerProject/Assets/Shared/GameSettings.sdgamesettings
