+++ XenkoLiveEditorContainerProject/Assets/Shared/MainScene.sdscene
