+++ XenkoLiveEditorContainerProject/Assets/Shared/Skybox.sdsky
