+++ XenkoLiveEditorContainerProject/Assets/Shared/Sphere Material.sdmat
