+++ XenkoLiveEditorContainerProject/Assets/Shared/Sphere.sdpromodel
