+++ XenkoLiveEditorContainerProject/XenkoLiveEditorContainerProject.Game/AutoAddEntitiesScript.cs
@@ -1,4 +1,4 @@
-using Xenko.Engine;
+using Stride.Engine;
 using System;
 using System.Collections.Generic;
 using System.Linq;
