+++ XenkoLiveEditorContainerProject/XenkoLiveEditorContainerProject.Game/TestScript.cs
@@ -1,5 +1,5 @@
 using Stride.Core.Mathematics;
-using Xenko.Engine;
+using Stride.Engine;
 
 namespace StrideLiveEditorContainerProject
 {
