+++ XenkoLiveEditorContainerProject/XenkoLiveEditorContainerProject.Windows/XenkoLiveEditorContainerProjectApp.cs
@@ -1,5 +1,5 @@
 using System;
-using Xenko.Engine;
+using Stride.Engine;
 using StrideLiveEditor;
 
 namespace StrideLiveEditorContainerProject
